helm-remove-defaults/pyproject.toml
[project]
name = "helm-remove-defaults"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pyyaml>=6.0.2",
    "ruamel-yaml>=0.18.10",
]
